requirements.txt
astroid==1.6.1
Django==1.11.8
django-cors-headers==2.1.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.0.0
PyJWT==1.5.3
pylint==1.8.2
pytz==2017.3
six==1.11.0
wrapt==1.10.11
